add tests for gs and minio
