+++ docker/compose/docker-compose.yaml
@@ -183,6 +183,7 @@ services:
       - APROC_INPUT_STORAGE_HEADERS=${APROC_INPUT_STORAGE_HEADERS}
       - APROC_INPUT_STORAGE_DOMAIN=${APROC_INPUT_STORAGE_DOMAIN}
       - APROC_INPUT_STORAGE_FORCE_DOWNLOAD=${APROC_INPUT_STORAGE_FORCE_DOWNLOAD}
+      - INGESTED_FOLDER=${INGESTED_FOLDER}
     volumes:
       - ${PWD}/test/inputs:/inputs:ro
       - ${PWD}/outbox/:/outbox:wr
@@ -253,7 +254,7 @@ services:
       - APROC_INPUT_STORAGE_HEADERS=${APROC_INPUT_STORAGE_HEADERS}
       - APROC_INPUT_STORAGE_DOMAIN=${APROC_INPUT_STORAGE_DOMAIN}
       - APROC_INPUT_STORAGE_FORCE_DOWNLOAD=${APROC_INPUT_STORAGE_FORCE_DOWNLOAD}
-
+      - INGESTED_FOLDER=${INGESTED_FOLDER}
     volumes:
       - ${PWD}/test/inputs:/inputs:ro
     ports:
