+++ python/aias_common/access/storages/s3.py
@@ -48,8 +48,7 @@ class S3Storage(AbstractStorage):
         return False
 
     def exists(self, href: str):
-        import botocore.exceptions
-
+        import botocore.exceptions       
         try:
             return self.__head_object(href)['ResponseMetadata']['HTTPStatusCode'] == 200
         except botocore.exceptions.ClientError:
@@ -95,13 +94,11 @@ class S3Storage(AbstractStorage):
 
     def __head_object(self, href: str):
         return self.get_storage_parameters()["client"].head_object(
-                Bucket=self.get_configuration().bucket,
-                Key=self.__get_href_key(href)
-            )
+            Bucket=self.get_configuration().bucket,
+            Key=self.__get_href_key(href))
 
     def is_file(self, href: str):
         import botocore.exceptions
-
         try:
             return self.__head_object(href)['ResponseMetadata']['HTTPStatusCode'] == 200
         except botocore.exceptions.ClientError:
@@ -148,8 +145,12 @@ class S3Storage(AbstractStorage):
 
         return files + dirs
 
-    def get_last_modification_time(self, href: str):
-        return self.__head_object(href)['LastModified'].timestamp()
+    def get_last_modification_time(self, href: str):        
+        import botocore.exceptions
+        try:
+            return self.__head_object(href)['LastModified'].timestamp()
+        except botocore.exceptions.ClientError:
+            return None
 
     def get_creation_time(self, href: str):
         # There is no difference in s3 between last update and creation date
