+++ test/env_fs.sh
@@ -1,0 +1,1 @@
+export INGESTED_FOLDER=/inputs
