+++ test/env_minio.sh
@@ -1,0 +1,1 @@
+export INGESTED_FOLDER=http://minio:9000
