+++ test/env_template.sh
@@ -96,8 +96,6 @@ export AGATE_LOGGER_LEVEL=DEBUG
 
 # FAM
 # To test FAM with Google Storage
-# export INGESTED_FOLDER=gs://gisaia-public/test-aias/
-export INGESTED_FOLDER=/inputs
 export FAM_LOGGER_LEVEL=DEBUG
 export FAM_CORS_ORIGINS="*"
 export FAM_CORS_METHODS="*"
