+++ test/tests_g1.sh
@@ -10,5 +10,3 @@ docker ps
 
 echo "run test.aproc_download_tests"
 docker run --rm -v `pwd`:/app/  --network compose_aias pythontests python3 -m test.aproc_download_tests
-
-./test/stop_stack.sh
