+++ test/tests_g3.sh
@@ -11,5 +11,3 @@ docker run --rm -v `pwd`:/app/  --network compose_aias pythontests python3 -m te
 
 echo "run test.aproc_dc3build_tests"
 docker run --rm -v `pwd`:/app/  --network compose_aias pythontests python3 -m test.aproc_dc3build_tests
-
-./test/stop_stack.sh
